Fix remaining functions after rebase
